Added SaveChangesAsync Interceptor

+++ Core/ECommerce.Domain/Entities/Product.cs
@@ -5,8 +5,8 @@ namespace ECommerce.Domain.Entities
     public class Product : BaseEntity
     {
         public string Name { get; set; }
-        public int Price { get; set; }
-        public long Stock { get; set; }
+        public double Price { get; set; }
+        public int Stock { get; set; }
         public ICollection<Order> Orders { get; set; }
     }
 }
